Create comsaphanaextendedbars-bundle.js
This file is included so you can easily look at the JavaScript that creates the chart. However, this code is for SAP Lumira 1.16 or so, and doesn't automatically work in newer versions. I would recommend you download the .profile file and load that into your VizPacker instead.The .profile file was adjusted and updated for SAP Lumira 1.21 and should work in newer versions as well.